shared types for Language and languages.
