+++ app/docs/posts/constants.ts
@@ -1,15 +1,3 @@
-export interface Language {
-  readonly name: string;
-  readonly value: string;
-}
-
-export const languages: readonly Language[] = [
-  { name: "cURL", value: "curl" },
-  { name: "Python", value: "python" },
-  { name: "JavaScript", value: "javascript" },
-  { name: "TypeScript", value: "typescript" },
-] as const;
-
 export const codeExamples = {
   curl: `curl -X GET "https://api.recoupable.com/api/posts" \\
   -H "Content-Type: application/json" \\
